Optimize Raspi Code
Optimize Raspi Compiler Flags and use float. Raspi Zero CPU works on
float precision so using doubles is actually slower

+++ jellED/core/beatDetection/include/filterStage.h
@@ -9,7 +9,7 @@ class FilterStage {
 public:
     FilterStage(){};
     virtual ~FilterStage(){}
-    virtual double apply(const double sample) = 0;
+    virtual float apply(const float sample) = 0;
 };
 
 } // namespace jellED
